CB-9926: Add retries to the legacy health check's ping
Add retries to the legacy health check's ping.

Refactor the health checks to allow it to be retry on each instance of
FreeIPA.
