+++ freeipa/src/test/java/com/sequenceiq/freeipa/flow/stack/start/OneFreeIpaReachableAttemptTest.java
@@ -21,7 +21,7 @@ import com.sequenceiq.freeipa.client.FreeIpaClientException;
 import com.sequenceiq.freeipa.client.model.RPCResponse;
 import com.sequenceiq.freeipa.entity.InstanceMetaData;
 import com.sequenceiq.freeipa.entity.Stack;
-import com.sequenceiq.freeipa.service.stack.FreeIpaHealthDetailsService;
+import com.sequenceiq.freeipa.service.stack.FreeIpaInstanceHealthDetailsService;
 
 @ExtendWith(MockitoExtension.class)
 class OneFreeIpaReachableAttemptTest {
@@ -29,7 +29,7 @@ class OneFreeIpaReachableAttemptTest {
     public static final long ID = 99927L;
 
     @Mock
-    private FreeIpaHealthDetailsService freeIpaHealthDetailsService;
+    private FreeIpaInstanceHealthDetailsService freeIpaInstanceHealthDetailsService;
 
     private OneFreeIpaReachableAttempt oneFreeIpaReachableAttemptUnderTest;
 
@@ -49,32 +49,32 @@ class OneFreeIpaReachableAttemptTest {
         InstanceMetaData instance2 = new InstanceMetaData();
         instanceSet = Set.of(instance1, instance2);
 
-        oneFreeIpaReachableAttemptUnderTest = new OneFreeIpaReachableAttempt(freeIpaHealthDetailsService, stack, instanceSet);
+        oneFreeIpaReachableAttemptUnderTest = new OneFreeIpaReachableAttempt(freeIpaInstanceHealthDetailsService, stack, instanceSet);
         response = new RPCResponse<>();
     }
 
     @Test
     public void testAttemptSucceed() throws Exception {
         response.setResult(Boolean.TRUE);
-        when(freeIpaHealthDetailsService.checkFreeIpaHealth(any(), any())).thenReturn(response);
+        when(freeIpaInstanceHealthDetailsService.checkFreeIpaHealth(any(), any())).thenReturn(response);
         Assertions.assertEquals(oneFreeIpaReachableAttemptUnderTest.process().getState(), AttemptState.FINISH);
-        verify(freeIpaHealthDetailsService, times(1)).checkFreeIpaHealth(any(), any());
+        verify(freeIpaInstanceHealthDetailsService, times(1)).checkFreeIpaHealth(any(), any());
     }
 
     @Test
     public void testAttemptFailed() throws Exception {
         response.setResult(Boolean.FALSE);
-        when(freeIpaHealthDetailsService.checkFreeIpaHealth(any(), any())).thenReturn(response);
+        when(freeIpaInstanceHealthDetailsService.checkFreeIpaHealth(any(), any())).thenReturn(response);
         Assertions.assertEquals(oneFreeIpaReachableAttemptUnderTest.process().getState(), AttemptState.CONTINUE);
-        verify(freeIpaHealthDetailsService, times(2)).checkFreeIpaHealth(any(), any());
+        verify(freeIpaInstanceHealthDetailsService, times(2)).checkFreeIpaHealth(any(), any());
     }
 
     @Test
     public void testAttemptFailedWithException() throws Exception {
         response.setResult(Boolean.FALSE);
-        when(freeIpaHealthDetailsService.checkFreeIpaHealth(any(), any())).thenThrow(FreeIpaClientException.class);
+        when(freeIpaInstanceHealthDetailsService.checkFreeIpaHealth(any(), any())).thenThrow(FreeIpaClientException.class);
         Assertions.assertEquals(oneFreeIpaReachableAttemptUnderTest.process().getState(), AttemptState.CONTINUE);
-        verify(freeIpaHealthDetailsService, times(2)).checkFreeIpaHealth(any(), any());
+        verify(freeIpaInstanceHealthDetailsService, times(2)).checkFreeIpaHealth(any(), any());
     }
 
     @Test
